requirements.txt
pandas
datasets
transformers
peft
accelerate
bitsandbytes
scipy
discord.py
python-dotenv
